+++ apps/nextjs-web/app/client/strapi-client.tsx
@@ -1,0 +1,11 @@
+import { registerUrql } from '@urql/next/rsc'
+import { cacheExchange, createClient, fetchExchange } from '@urql/core'
+
+const makeClient = () => {
+  return createClient({
+    url: 'http://localhost:3000/graphql',
+    exchanges: [cacheExchange, fetchExchange],
+  })
+}
+
+export const { getClient } = registerUrql(makeClient)
